requirements.txt
streamlit
pandas
plotly
python-dotenv
groq
